requirements-base.txt
requests>=2.9.1
oauthlib>=1.0.3
requests-oauthlib>=0.6.1
six>=1.10.0
PyJWT>=1.4.0
python-jose==3.0.1
